[PLAYER-4808] Moved progress bar to new common folder and volume panel updated
