+++ sdk/react/bottomOverlay.js
@@ -29,7 +29,7 @@ const AndroidAccessibility = NativeModules.AndroidAccessibility;
 const AccessibilityUtils = require('./accessibilityUtils');
 const Log = require('./log');
 const Utils = require('./utils');
-const ProgressBar = require('./progressBar');
+const ProgressBar = require('./common/progressBar');
 const ControlBar = require('./controlBar');
 const ResponsiveDesignManager = require('./responsiveDesignManager');
 const styles = Utils.getStyles(require('./style/bottomOverlayStyles.json'));
