+++ sdk/react/common/style/progressBarStyles.json
