+++ sdk/react/panels/VolumePanel.js
@@ -8,11 +8,12 @@ import {
   Slider,
   Text,
   TouchableHighlight,
-  View,
+  View
 } from 'react-native';
 
 import {
   BUTTON_NAMES,
+  UI_SIZES,
   CELL_TYPES
 } from '../constants';
 
@@ -21,12 +22,14 @@ const styles = Utils.getStyles(require('./style/VolumePanelStyles'));
 
 const constants = {
   animationDuration: 1000
- }
+}
 
 class VolumePanel extends React.Component {
   static propTypes = {
     onDismiss: PropTypes.func,
-    config: PropTypes.object
+    volume: PropTypes.number.isRequired,
+    width: PropTypes.number.isRequired,
+    config: PropTypes.object,
   };
 
   state = {
@@ -40,26 +43,45 @@ class VolumePanel extends React.Component {
         this.state.opacity,
         {
           toValue: 1,
-          duration: constants.animationDuration,
-          delay: 0
+          duration: constants.animationDuration
         }),
     ]).start();
   };
 
   _renderVolumeIcon= () => {
+    const iconConfig = (this.props.volume > 0) ? this.props.config.icons.volume : this.props.config.icons.volumeOff;
+    const fontFamilyStyle = {fontFamily: iconConfig.fontFamilyName};
 
+    return (
+      <View style={styles.volumeIconContainer}>
+        <Text style={[styles.volumeIcon, fontFamilyStyle]}>
+          {iconConfig.fontString}
+        </Text>
+      </View>
+    );
   };
 
-  _renderVolumeIcon= () => {
-
+  _renderVolumeSlider= () => {
+    return (
+      <View style={styles.sliderContainer}>
+        <Slider
+          step={1}
+          maximumValue={100}
+          // onValueChange={this.change.bind(this)}
+          value={40}
+        /> 
+      </View>
+    );
   };
+
 
   render() {
-    const animationStyle = {opacity:this.state.opacity};
+    const animationStyle = {opacity: this.state.opacity};
 
     return (
       <Animated.View style={[styles.container, animationStyle]}>
-
+        {this._renderVolumeIcon()}
+        {this._renderVolumeSlider()}
       </Animated.View>
     );
   }
